requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
motor
pymongo
pdf2image
python-multipart
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
boto3
PyJWT
bcrypt
slowapi
httpx
sib-api-v3-sdk
jinja2
psutil
openpyxl